* Use vcpkg insted of hunter

* Add merkle module

* Add lightnode module

* Merge changes from release-3.0.0-rc4

* Update cxx standard to c++20

* Merge bcos-utilities bcos-crypto bcos-boostssl to this repo

+++ bcos-codec/abi/ContractABICodec.h
@@ -243,7 +243,7 @@ struct Offset<T, List...>
 {
     enum
     {
-        value = Offset<T>::value + Offset<List...>::value
+        value = (size_t)Offset<T>::value + (size_t)Offset<List...>::value
     };
 };
